+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG for Sulu
 
 * dev-develop
     * ENHANCEMENT #3154 [All]                 Upgrade symfony to ^3.0
+    * ENHANCEMENT #3266 [ContentBundle]         Added locale parameter to teaser-selection-list
 
 * 1.5.1 (2017-03-16)
     * ENHANCEMENT #3256 [DocumentManagerBundle] Added VersionNotFoundException to fos_rest configuration